Rename ProtocolUtils -> callbackUrlUtils
